chore(lotto): move Lotto.js file and rename lottoFormatUtils.js
- rename to resolve case sensitivity issue with lottoFormatUtils import
