+++ src/utils/lottoFormatUtils.js
